timeit: script to extract timing information from shortlog
Also tee the timing information on the job output

this patch adds a script, timeit.py. This python script reads the
mkcloud-shortlog generated for each mkcloud job and looks for specific
logs and using regex parses out the start and finish times.
Then spits out the collected information and duration.